requirements.txt
Pillow
requests
func-timeout
python-dotenv